[291] refactored naming scheme and created final dict
