+++ src/nlp_service/util/models/common_example_model.py
@@ -1,4 +1,4 @@
-class IntentModel:
+class CommonExampleModel:
     def __init__(self):
         self.intent_dict = {
             "text": "",
