--- src/nlp_service/util/models/json_model.py
@@ -1,11 +1,0 @@
-class JsonModel:
-    def __init__(self):
-        self.nlu_dict = {"rasa_nlu_data": {
-            "regex_features": [
-
-            ],
-            "entity_synonyms": "",
-            "common_examples": [
-
-            ]}
-        }
