+++ src/nlp_service/util/parse_dataset_test.py
@@ -1,5 +1,5 @@
 import unittest
-from util.create_json import CreateJson
+from util.parse_dataset import CreateJson
 
 
 class TestMain(unittest.TestCase):
